refactoring the migration command
